pyproject.toml
[project]
name = "llm4se-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.2",
    "pydantic>=2.12.5",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "sentence-transformers>=5.2.3",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
